requirements.txt
GitPython
pyGithub
python-gitlab
